requirements.txt

asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0    
cloudinary==1.35.0
colorama==0.4.6
decorator==4.4.2
dj3-cloudinary-storage==0.0.6
Django==4.2.5
django-jsonfield==1.4.1      
djangorestframework==3.14.0  
djangorestframework-simplejwt==5.3.0
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mysqlclient==2.2.0
numpy==1.26.0
Pillow==10.0.1
proglog==0.1.10
pydub==0.25.1
PyJWT==2.8.0
python-magic==0.4.27
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16